[SC-65] Support ConnectionURI in multiple SC portfolios
To support ConnectionURI in multiple SC portfolios we need to set the
tag Protected/AccessApprovedCaller with the logged in user's
SCEndUser_ROLE_ID:Synapse_User_ID.  The ROLE_ID can either be
SCEndUser or SCExternalEndUser based on which synapse team the user is
on.  This change will dynmically lookup the correct value to assign to the
Protected/AccessApprovedCaller tag.