[SC-409] Upgrade from IMDSv1 to IMDSv2
Testing the Amazon Linux 2023 image in scipool-dev is failing because
the EC2 instance created by Service Catalog requires IMDSv2, and an
empty response from IMDSv1 is causing this script to fail.

Follow the user guide [1] to make a request to IMDSv2 instead, and
rely on the PATH environment variable for looking up the location of
executables since Amazon Linux pre-installs aws-cli under `/usr/bin`.

[1] https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instancedata-data-retrieval.html